+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ## v1.5.0
 
-*2024-11-11*
+*2024-11-19*
 
 **Added**
 
@@ -8,6 +8,7 @@
 - Reference provider for Wikitext now also includes HTML tags and wiki links
 - Rename provider for Wikitext, including parser functions, templates, arguments and wiki links
 - Completion for tag attribute keys in Wikitext
+- Folding range provider for Wikitext
 
 **Fixed**